removed constants.js from Table
