--- app/javascript/src/components/Dashboard/Contacts/Table/constants.js
@@ -1,14 +1,0 @@
-export const CONTACTS_UNIQUE_DATA = [
-  {
-    name: "Ronald Richards",
-    role: "owner",
-    email: "ronaldrichards@gmail.com",
-    created_at: "2020-01-13T05:07:55.886Z",
-  },
-  {
-    name: "Jacob Jones",
-    role: "owner",
-    email: "jacob@gmail.com",
-    created_at: "2022-09-13T05:07:55.886Z",
-  },
-];
