+++ app/javascript/src/components/Dashboard/Contacts/constants.js
@@ -1,5 +1,20 @@
 import * as yup from "yup";
 
+export const CONTACTS_UNIQUE_DATA = [
+  {
+    name: "Ronald Richards",
+    role: "owner",
+    email: "ronaldrichards@gmail.com",
+    created_at: "2020-01-13T05:07:55.886Z",
+  },
+  {
+    name: "Jacob Jones",
+    role: "owner",
+    email: "jacob@gmail.com",
+    created_at: "2022-09-13T05:07:55.886Z",
+  },
+];
+
 export const DUMMY_ROLES = [
   {
     label: "Owner",
